Refactor otter-trace code into multiple smaller source files, focussing on one specific abstraction per file
