+++ include/otter/trace-structs.h
@@ -3,6 +3,7 @@
 
 #include <stdint.h>
 #include <stdbool.h>
+#include <pthread.h>
 #include <otf2/otf2.h>
 
 #include "otter/otter-ompt-header.h"
@@ -10,6 +11,12 @@
 #include "otter/queue.h"
 #include "otter/stack.h"
 #include "otter/trace.h"
+#include "otter/trace-location.h"
+#include "otter/trace-region-parallel.h"
+#include "otter/trace-region-workshare.h"
+#include "otter/trace-region-master.h"
+#include "otter/trace-region-sync.h"
+#include "otter/trace-region-task.h"
 
 /* Forward definitions */
 typedef struct trace_parallel_region_attr_t trace_parallel_region_attr_t;
@@ -33,7 +40,7 @@ struct trace_parallel_region_attr_t {
 
 /* Attributes of a workshare region */
 struct trace_wshare_region_attr_t {
-    ompt_work_t     type;
+    ompt_work_t     type;       // TODO: decouple
     uint64_t        count;
 };
 
@@ -44,19 +51,19 @@ struct trace_master_region_attr_t {
 
 /* Attributes of a sync region */
 struct trace_sync_region_attr_t {
-    ompt_sync_region_t  type;
+    ompt_sync_region_t  type;       // TODO: decouple
     unique_id_t         encountering_task_id;
 };
 
 /* Attributes of a task region */
 struct trace_task_region_attr_t {
     unique_id_t         id;
-    ompt_task_flag_t    type;
-    ompt_task_flag_t    flags;
+    ompt_task_flag_t    type;               // TODO: decouple
+    ompt_task_flag_t    flags;              // TODO: decouple
     int                 has_dependences;
     unique_id_t         parent_id;
-    ompt_task_flag_t    parent_type;
-    ompt_task_status_t  task_status;
+    ompt_task_flag_t    parent_type;        // TODO: decouple
+    ompt_task_status_t  task_status;        // TODO: decouple
 };
 
 typedef union {
@@ -78,73 +85,6 @@ struct trace_region_def_t {
     otter_stack_t       *rgn_stack;
     trace_region_attr_t  attr;    
 };
-
-/* Store values needed to register location definition (threads) with OTF2 */
-struct trace_location_def_t {
-    unique_id_t             id;
-    ompt_thread_t           thread_type;
-    uint64_t                events;
-    otter_stack_t                *rgn_stack;
-    otter_queue_t                *rgn_defs;
-    otter_stack_t                *rgn_defs_stack;
-    OTF2_LocationRef        ref;
-    OTF2_LocationType       type;
-    OTF2_LocationGroupRef   location_group;
-    OTF2_AttributeList     *attributes;
-    OTF2_EvtWriter         *evt_writer;
-    OTF2_DefWriter         *def_writer;
-};
-
-/* Create new location */
-trace_location_def_t *
-trace_new_location_definition(
-    uint64_t              id,
-    ompt_thread_t         thread_type,
-    OTF2_LocationType     loc_type, 
-    OTF2_LocationGroupRef loc_grp);
-
-/* Create new region */
-trace_region_def_t *
-trace_new_parallel_region(
-    unique_id_t    id, 
-    unique_id_t    master,
-    unique_id_t    encountering_task_id,
-    int            flags,
-    unsigned int   requested_parallelism);
-
-trace_region_def_t *
-trace_new_workshare_region(
-    trace_location_def_t *loc,
-    ompt_work_t           wstype,
-    uint64_t              count,
-    unique_id_t           encountering_task_id);
-
-trace_region_def_t *
-trace_new_master_region(
-    trace_location_def_t *loc,
-    unique_id_t           encountering_task_id);
-
-trace_region_def_t *
-trace_new_sync_region(
-    trace_location_def_t *loc,
-    ompt_sync_region_t    stype,
-    unique_id_t           encountering_task_id);
-
-trace_region_def_t *
-trace_new_task_region(
-    trace_location_def_t *loc,
-    trace_region_def_t   *parent_task_region,
-    unique_id_t           task_id,
-    ompt_task_flag_t      flags,
-    int                   has_dependences);
-
-/* Destroy location/region */
-void trace_destroy_location(trace_location_def_t *loc);
-void trace_destroy_parallel_region(trace_region_def_t *rgn);
-void trace_destroy_workshare_region(trace_region_def_t *rgn);
-void trace_destroy_master_region(trace_region_def_t *rgn);
-void trace_destroy_sync_region(trace_region_def_t *rgn);
-void trace_destroy_task_region(trace_region_def_t *rgn);
 
 /* pretty-print region definitions */
 void trace_region_pprint(FILE *fp, trace_region_def_t *r, const char func[], const int line);
